unsigned

+++ i2c_descriptor.hpp
@@ -4,5 +4,5 @@ class i2c_descriptor:file_descriptor
 {
 public:
   i2c_descriptor(const char device[],const int address);
-  int read_bytes(const char send[],char buf[],size_t buf_size)const;
+  int read_bytes(const char send[],char buf[],unsigned buf_size)const;
 };
